Fix #380
