+++ Lib/Collectors/BaseCollector.cs
@@ -1,14 +1,17 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
+using AttackSurfaceAnalyzer.Objects;
 using AttackSurfaceAnalyzer.Types;
 using AttackSurfaceAnalyzer.Utils;
 using Serilog;
 using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Globalization;
 using System.Linq;
 using System.Threading;
+using System.Threading.Tasks;
 
 namespace AttackSurfaceAnalyzer.Collectors
 {
@@ -23,80 +26,18 @@ namespace AttackSurfaceAnalyzer.Collectors
 
         private readonly int _numCollected = 0;
 
+        public ConcurrentQueue<CollectObject> Results { get; } = new ConcurrentQueue<CollectObject>();
+
         public void Execute()
         {
             if (!CanRunOnPlatform())
             {
                 Log.Warning(Strings.Get("Err_PlatIncompat"), GetType().ToString());
-                return;
             }
-            Start();
-
-            DatabaseManager.BeginTransaction();
-
-            var StopWatch = Stopwatch.StartNew();
-
-            ExecuteInternal();
-
-            StopWatch.Stop();
-            TimeSpan t = TimeSpan.FromMilliseconds(StopWatch.ElapsedMilliseconds);
-            string answer = string.Format(CultureInfo.InvariantCulture, "{0:D2}h:{1:D2}m:{2:D2}s:{3:D3}ms",
-                                    t.Hours,
-                                    t.Minutes,
-                                    t.Seconds,
-                                    t.Milliseconds);
-            Log.Debug(Strings.Get("Completed"), GetType().Name, answer);
-
-            var prevFlush = DatabaseManager.Connections.Select(x => x.WriteQueue.Count).Sum();
-            var totFlush = prevFlush;
-
-            var printInterval = 10;
-            var currentInterval = 0;
-
-            StopWatch = Stopwatch.StartNew();
-
-            while (DatabaseManager.HasElements())
+            else
             {
-                Thread.Sleep(1000);
-
-                if (currentInterval++ % printInterval == 0)
-                {
-                    var actualDuration = (currentInterval < printInterval) ? currentInterval : printInterval;
-                    var sample = DatabaseManager.Connections.Select(x => x.WriteQueue.Count).Sum();
-                    var curRate = prevFlush - sample;
-                    var totRate = (double)(totFlush - sample) / StopWatch.ElapsedMilliseconds;
-                    try
-                    {
-                        t = (curRate > 0) ? TimeSpan.FromMilliseconds(sample / ((double)curRate / (actualDuration * 1000))) : TimeSpan.FromMilliseconds(99999999); //lgtm[cs/loss-of-precision]
-                        answer = string.Format(CultureInfo.InvariantCulture, "{0:D2}h:{1:D2}m:{2:D2}s:{3:D3}ms",
-                                                t.Hours,
-                                                t.Minutes,
-                                                t.Seconds,
-                                                t.Milliseconds);
-                        Log.Debug("Flushing {0} results. ({1}/{4}s {2:0.00}/s overall {3} ETA)", sample, curRate, totRate * 1000, answer, actualDuration);
-                    }
-                    catch (Exception e) when (
-                        e is OverflowException)
-                    {
-                        Log.Debug($"Overflowed: {curRate} {totRate} {sample} {t} {answer}");
-                        Log.Debug("Flushing {0} results. ({1}/s {2:0.00}/s)", sample, curRate, totRate * 1000);
-                    }
-                    prevFlush = sample;
-                }
-
+                ExecuteInternal();
             }
-
-            StopWatch.Stop();
-            t = TimeSpan.FromMilliseconds(StopWatch.ElapsedMilliseconds);
-            answer = string.Format(CultureInfo.InvariantCulture, "{0:D2}h:{1:D2}m:{2:D2}s:{3:D3}ms",
-                                    t.Hours,
-                                    t.Minutes,
-                                    t.Seconds,
-                                    t.Milliseconds);
-            Log.Debug("Completed flushing in {0}", answer);
-
-            DatabaseManager.Commit();
-            Stop();
         }
 
         public abstract bool CanRunOnPlatform();
@@ -114,7 +55,6 @@ namespace AttackSurfaceAnalyzer.Collectors
         {
             _running = RUN_STATUS.RUNNING;
             watch = System.Diagnostics.Stopwatch.StartNew();
-
             Log.Information(Strings.Get("Starting"), GetType().Name);
         }
